DDF-1943: Create non-static Security class in org.codide.ddf.common.util
Created new, non-static Security class in org.codice.ddf.common to make unit testing
easier, added a new runWithSubjectOrElevate() method, updated old Security class in
ddf.security.common.util to delegate to the new one and updated and moved unit tests.
